requirements.txt
streamlit>=1.34
langchain>=0.1.9
langchain-google-genai>=2.1.8
pytesseract>=0.3.10
pdfplumber>=0.10.2
pillow>=10.0.0
python-docx>=1.0.0
docx2txt>=0.8
pdf2image>=1.16.3
python-dotenv>=1.0.0